Quantization (#189)
 - Quantize network when loading into engine
 - Remove all additional information from network file except hash 

http://chess.grantnet.us/test/19994/

ELO   | 23.65 +- 10.15 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | 3.00 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 2192 W: 607 L: 458 D: 1127

http://chess.grantnet.us/test/19995/

ELO   | 33.63 +- 11.36 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=64MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 1368 W: 327 L: 195 D: 846

Bench: 2142286

+++ src/network.h
@@ -32,9 +32,9 @@ public:
         hidden_neurons.back().fill(0.0f);
     }
 
-    float calculate_last_layer();
+    int32_t calculate_last_layer();
 
-    float feed(NetworkInput const &);
+    int32_t feed(NetworkInput const &);
 
     void update_hidden_layer(NetworkUpdateList const &);
 
@@ -55,11 +55,11 @@ public:
     }
 
 private:
-    std::vector<std::array<float, HIDDEN_SIZE>> hidden_neurons;
+    std::vector<std::array<int32_t, HIDDEN_SIZE>> hidden_neurons;
 
-    static std::array<std::array<float, HIDDEN_SIZE>, INPUT_SIZE> hidden_weights;
-    static std::array<float, HIDDEN_SIZE> hidden_biases;
-    static std::array<float, HIDDEN_SIZE> output_weights;
-    static float output_bias;
+    static std::array<std::array<int32_t, HIDDEN_SIZE>, INPUT_SIZE> hidden_weights;
+    static std::array<int32_t, HIDDEN_SIZE> hidden_biases;
+    static std::array<int32_t, HIDDEN_SIZE> output_weights;
+    static int32_t output_bias;
     static std::uint32_t hash;
 };
